requirements.txt
Flask
psycopg2-binary
gunicorn
# boto3 is removed as we are no longer using AWS S3
